fill out stats class
